apply_chat_template: optionally take tools from example
In order to apply ORPO fine tuning against a model's use of tools, it
is necessary to provide the chat template with the set of tools
available to the model so the prompts can be consistent with those
used at inference. Extend apply_chat_template in data_utils.py to take
a list of JSON Schema tool dictionaries from the example if no tools
are provided in the call.